+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## _(in-progress)_
+* New: Add setting for configuring time format
+    * New: Add "Blog Time Format" setting to the "General Settings" page.
+    * New: Add link to setting from plugin's action links
+    * New: Add `initialize_setting()`, `allowed_options()`, `display_option()`, and `plugin_action_links()`
 * Change: Use default time format if parameter or filter attempts to configure an empty string or non-string value
 * Fix: Add support for the 'T' timezone format character to the dynamic clock (support for which was removed from Moment.js awhile ago)
 * Change: Note compatibility through WP 5.7+